PR for profile_speedups
