+++ bootleg/extract_all_entities.py
@@ -150,8 +150,8 @@ def run_model(config, run_config_path=None):
         load_dir=os.path.join(
             config.data_config.entity_dir, config.data_config.entity_map_dir
         ),
-        alias_cand_map_fld=config.data_config.alias_cand_map,
-        alias_idx_fld=config.data_config.alias_idx_map,
+        alias_cand_map_dir=config.data_config.alias_cand_map,
+        alias_idx_dir=config.data_config.alias_idx_map,
     )
     qid2eid = entity_symbols.get_qid2eid_dict()
     eid2qid = {v: k for k, v in qid2eid.items()}
